Wrap the navbar in an <md-content> as required; fix a typo
md-nav-bar's apparently need to be wrapped in an <md-content> tag,
which allows you to set the theme, class, and CSS style of it.
I went ahead and had it set to the background color Ariel was using.
I also added a {{currentNavItem}} as a demonstration, you can take it out.
And, finally, I wrapped all this in the top-level <div> so it is seamless.