I am VICTORIOUS
So the error ended up being that first of all, I was accidentally
mistaking rgb(255,255,255) for being black. But I was setting
everything to white. I was also never setting the start event to work.
And in my start event I was already calling step, so I couldn’t get it
to start only on hover.
But by removing step from within the start function, I was able to
choose when to call and it and when not to.

Basically, now it lights up in perfect sequence, AND it only starts
when I hover.