Used Boolean logic to downsize code
I think I understand how if statements evaluate a bit better, so I
tinkered with the code to see if I understood the logic of each section
better. So this is me downsizing code based on my understanding! Please
let me know if I did something wrong. But it seems to be working as
intended